Move zod types
