+++ projects/plugins/boost/app/assets/src/js/modules/image-size-analysis/store/isa-data.ts
@@ -1,67 +1,7 @@
 import { derived } from 'svelte/store';
 import { z } from 'zod';
 import { jetpack_boost_ds } from '../../../stores/data-sync-client';
-import { getPreloadingImages } from './preloading-image';
-
-/**
- * Zod Types
- */
-const Dimensions = z.object( {
-	width: z.number(),
-	height: z.number(),
-} );
-
-const ImageData = z.object( {
-	id: z.string(),
-	thumbnail: z.string(),
-	image: z.object( {
-		url: z.string(),
-		dimensions: z.object( {
-			file: Dimensions,
-			expected: Dimensions,
-			size_on_screen: Dimensions,
-		} ),
-		weight: z.object( {
-			current: z.number(),
-			potential: z.number(),
-		} ),
-	} ),
-	page: z.object( {
-		id: z.number(),
-		url: z.string(),
-		title: z.string(),
-	} ),
-	edit_url: z.string().url(),
-	device_type: z.enum( [ 'phone', 'desktop' ] ),
-	instructions: z.string(),
-} );
-
-const ImageSizeAnalysis = z
-	.object( {
-		query: z.object( {
-			page: z.number(),
-			group: z.string(),
-			search: z.string(),
-		} ),
-		data: z.object( {
-			last_updated: z.number(),
-			total_pages: z.number(),
-			images: z.array( ImageData ),
-		} ),
-	} )
-	// Prevent fatal error when this module isn't available.
-	.catch( {
-		query: {
-			page: 0,
-			group: '',
-			search: '',
-		},
-		data: {
-			last_updated: 0,
-			total_pages: 0,
-			images: getPreloadingImages( 10 ),
-		},
-	} );
+import { ImageData, ImageSizeAnalysis } from './zod-types';
 
 /**
  * Initialize the stores
