+++ projects/plugins/boost/app/assets/src/js/modules/image-size-analysis/store/zod-types.ts
@@ -1,0 +1,61 @@
+import { z } from 'zod';
+import { getPreloadingImages } from './preloading-image';
+/**
+ * Zod Types
+ */
+const Dimensions = z.object( {
+	width: z.number(),
+	height: z.number(),
+} );
+
+export const ImageData = z.object( {
+	id: z.string(),
+	thumbnail: z.string(),
+	image: z.object( {
+		url: z.string(),
+		dimensions: z.object( {
+			file: Dimensions,
+			expected: Dimensions,
+			size_on_screen: Dimensions,
+		} ),
+		weight: z.object( {
+			current: z.number(),
+			potential: z.number(),
+		} ),
+	} ),
+	page: z.object( {
+		id: z.number(),
+		url: z.string(),
+		title: z.string(),
+	} ),
+	edit_url: z.string().url(),
+	device_type: z.enum( [ 'phone', 'desktop' ] ),
+	instructions: z.string(),
+} );
+
+export const ImageSizeAnalysis = z
+	.object( {
+		query: z.object( {
+			page: z.number(),
+			group: z.string(),
+			search: z.string(),
+		} ),
+		data: z.object( {
+			last_updated: z.number(),
+			total_pages: z.number(),
+			images: z.array( ImageData ),
+		} ),
+	} )
+	// Prevent fatal error when this module isn't available.
+	.catch( {
+		query: {
+			page: 0,
+			group: '',
+			search: '',
+		},
+		data: {
+			last_updated: 0,
+			total_pages: 0,
+			images: getPreloadingImages( 10 ),
+		},
+	} );
